REFACTOR:
Integrated the generalised AST type "Node" throughout the parser.

This allows for Exprs to be treated as top-level nodes as well as Stmts, removing the need for 'wrapper' stmts such as "ExprStmt" & "FuncExprStmt".

This will also allow for the interpreter to rely on recursion more readily as funcs can take Nodes rather than explicitly Exprs or Stmts when the need arises.