Injecting Modules properly implemented:
Modules are found and instantiated at the start of the program. A dict is created with <string name, IModule instance>.

This dict is then passed into the Environment when created, becoming an Env 'Modules'.

When import statements are interpreted, they use 'GetModule' with the import name to find the IModlue from the Modules dict, then store that to the Env's variables dict as <string alias, IModule module>.

This means that while all modules are loaded at runtime, only those that the game author ha imported will bcome accessable / 'enter the gamestate', giving proper control over what ends up in the namespace and how.

Cleanup:
Cleaned 'using's across all files.