Integrated 'triggered' overlays, allowed GoTo's to take Variables, added 'run' & 'exit' keywords & appropriate tokens & statements.
Triggered Overlays: allows for things like a battle screen to be triggered without taking the player out of a scene.

Overlays are 'triggered' using the 'run' keyword. This works much like a GoTo, but for overlays instead. Appropriate flagging & storage has been added to Environment.

Triggered overlays can be exited by either an 'exit' statement, or a GoTo. An 'exit' statement will simply return the player to the scene they were previously in. As of current, as the intreactables & commands are repopulate from the old load, this means that an 'exit' will not update the state of available choices, if say for instance an if statement was conditioning the presentation of certain options to the player.

GoTo's & allowing GoTo's to take Variable Expressions allows for things like the following:

1. The player enters a scene with a battle in it.

2. 'back' is set to this scene's name.

3. The user is sent to the "battle" overlay.

4. Upon completion of the battle, a '-> back' is used to return the user to the scene they came from & to reload the scene itself. This provides a way to make reusable locations that can be entered from several scenes & exited correctly while still triggering a scene transition, thus allowing for updates to the gamestate to be detected. For example, presenting one set of texts to the user before triggering a battle if the battle has notyet been ran, then displaying a different set of messages & interactables afterwards.

In this way, scenes become evolving locations whose content can be reanalysed upon repeated visits, rather than static beats in a more formally directed story structure.