Polished the Battle Library. Updated Interpreter to allow for the creation of loops that don't require user input.
If a scene has no choices or commands in Local scopes & the GoTo flag has been triggered, the input section is skipped, and the GoTo is automatically triggered.

This allows for the creation of loops that automatically run.

It also allows for things like the result of a battle to automatically send the player to the next scene upon completion.