Removed CommandStmts & Added more complex NounStmt parsing s.t. complex user inputs can be created.
Got the foundations in to identify nouns & verbs & split things up & run. There are a few early return fails added but are some spots where I need further testing to ensure graceful failures so that the game doesn't crash on the user and they get an idea of what they need to do.