Added save & load functionality to the Environment s.t. a player can save & restore their progress.
Newtonsoft.Json used to serialize the Env into a json to save & then deserialize from the selected json to restore the gamestate.

The last GoTo location is sotred to an inaccessible value in the Env dictionary. When a save is loaded, this is used to set & flag the GoTo s.t. the player is properly moved to the correct scene by the game engine.

===========
Other:

Updated the PUML diagram

Updated Somna sample game to properly interact with the new save // load system

Moved table components (the table object & table row object) into the runtime in their own directory to better reflect when they are used.