Touchups:
Removed old TODO comments.
Deleted 'BasePlayer' module & the Somna implementation using it.
Modified Dice to return a float - originally returned an int, which would be treated as a string and require (num) conversion.
Interpreter: removed the 'silent scene' check as it was misplaced and would never work - it required knowledge of the results of interpretation before interpretation took place.